test footer-links
